+++ src/functionalTest/groovy/com/autonomousapps/android/ResSpec.groovy
@@ -1,15 +1,15 @@
 package com.autonomousapps.android
 
 import com.autonomousapps.android.projects.AndroidResourceProject
+import com.autonomousapps.android.projects.AttrResProject
 import org.gradle.testkit.runner.TaskOutcome
-import spock.lang.Unroll
 
 import static com.autonomousapps.utils.Runner.build
+import static com.google.common.truth.Truth.assertThat
 
 @SuppressWarnings("GroovyAssignabilityCheck")
 final class ResSpec extends AbstractAndroidSpec {
 
-  @Unroll
   def "plugin accounts for android resource usage (#gradleVersion AGP #agpVersion)"() {
     given:
     def project = new AndroidResourceProject(agpVersion)
@@ -28,5 +28,20 @@ final class ResSpec extends AbstractAndroidSpec {
 
     where:
     [gradleVersion, agpVersion] << gradleAgpMatrix()
+  }
+
+  def "detects attr usage in res file (#gradleVersion AGP #agpVersion)"() {
+    given:
+    def project = new AttrResProject(agpVersion)
+    gradleProject = project.gradleProject
+
+    when:
+    build(gradleVersion, gradleProject.rootDir, 'buildHealth')
+
+    then:
+    assertThat(project.expectedBuildHealth).containsExactlyElementsIn(project.actualBuildHealth())
+
+    where:
+    [gradleVersion, agpVersion] << gradleAgpMatrix(AGP_4_2)
   }
 }