+++ src/main/kotlin/com/autonomousapps/internal/utils/xml.kt
@@ -2,7 +2,10 @@ package com.autonomousapps.internal.utils
 
 import org.w3c.dom.Document
 import java.io.File
+import java.nio.file.Path
 import javax.xml.parsers.DocumentBuilderFactory
+
+internal fun buildDocument(path: Path): Document = buildDocument(path.toFile())
 
 internal fun buildDocument(file: File): Document = DocumentBuilderFactory.newInstance()
   .newDocumentBuilder()